Fix simple Gradio case and add Dockerfile (#38)
The current version of the pretrained models folder does not follow the
same folder structure expected by the defaults in the Gradio script
`interface.py` so since this is the simple / quick and easy path I
adjusted the script to follow the model folder structure. I also renamed
the default model in the script to match what's provided in the
pretrained models.

I also removed some parameters from the examples. `--image_size=1024` is
the default anyway, and `--model_path` also has a perfectly good default
value now so I removed it for clarity.

I also added a sample Dockerfile for a Gradio runtime environment :)

Finally, I clarified where you need to place the pretrained models
relative to everything else in order for the Gradio script to run
properly!